+++ common/src/main/java/org/fuin/objects4j/common/HasPublicStaticIsValidMethodValidator.java
@@ -2,10 +2,14 @@ package org.fuin.objects4j.common;
 
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
+import org.fuin.utils4j.Utils4J;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
+import java.util.Locale;
+import java.util.Map;
+import java.util.ResourceBundle;
 import java.util.function.Function;
 
 /**
@@ -15,6 +19,12 @@ import java.util.function.Function;
  * defined in the annotation and the return type of that method is {@literal boolean}.
  */
 public class HasPublicStaticIsValidMethodValidator implements ConstraintValidator<HasPublicStaticIsValidMethod, Object> {
+
+    private static final String PREFIX = HasPublicStaticIsValidMethod.class.getName() + ".";
+
+    private static final String POSTFIX = ".message";
+
+    private static final ResourceBundle MESSAGES = ResourceBundle.getBundle("ValidationMessages", Locale.getDefault());
 
     private String methodName;
 
@@ -46,21 +56,25 @@ public class HasPublicStaticIsValidMethodValidator implements ConstraintValidato
             final Method method = clasz.getMethod(methodName, paramClass);
             final int modifiers = method.getModifiers();
             if (!Modifier.isStatic(modifiers)) {
-                return new Result("Method '" + methodName + "' is not static (#1)", null);
+                return new Result(msg("1", Map.of("method", methodName)), null);
             }
             if (method.getReturnType() != boolean.class) {
-                return new Result("Method '" + methodName + "' does not return '" + clasz.getName() + "', but: " + method.getReturnType().getName() + " (#3)", null);
+                return new Result(msg("3", Map.of("method", methodName, "expectedType", clasz.getName(), "actualType", method.getReturnType().getName())), null);
             }
             final boolean valid = (boolean) method.invoke(clasz, (Object) null);
             if (!valid) {
-                return new Result("Method '" + methodName + "' is expected to return 'true' on 'null' argument, but was 'false' (#4)", null);
+                return new Result(msg("4", Map.of("method", methodName)), null);
             }
             return new Result(null, method);
         } catch (final NoSuchMethodException ex) {
-            return new Result("The method '" + methodName + "' is undefined or it is not public (#2)", null);
+            return new Result(msg("2", Map.of("method", methodName)), null);
         } catch (final IllegalAccessException | InvocationTargetException ex) {
             throw new IllegalStateException("Failed to execute method", ex);
         }
+    }
+
+    private static String msg(String key, Map<String, String> vars) {
+        return Utils4J.replaceVars(MESSAGES.getString(PREFIX + key + POSTFIX), vars);
     }
 
     /**