+++ common/src/test/java/org/fuin/objects4j/common/ArchitectureTest.java
@@ -41,7 +41,8 @@ class ArchitectureTest {
             .onlyDependOnClassesThat()
             .resideInAnyPackage(COMMON_PACKAGE,
                     "java.lang..", "java.util..", "java.io..", "java.text..",
-                    "jakarta.validation..", "jakarta.annotation..", "jakarta.enterprise.inject..");
+                    "jakarta.validation..", "jakarta.annotation..", "jakarta.enterprise.inject..",
+                    "org.fuin.utils4j..");
 
 
 }