+++ release-notes.md
@@ -2,7 +2,8 @@
 
 ## 0.11.1
 - Added new [TypeConstantValidator](common/src/main/java/org/fuin/objects4j/common/TypeConstantValidator.java) base class for validators based on existence a static constant in a class.
-- 
+- Added missing [ValidationMessages.properties](common/src/main/resources/ValidationMessages.properties)
+
 ## 0.11.0
 - Removed unnecessary Jackson serializer/deserializer classes
 - Added new [Objects4JacksonUtils](jackson/src/main/java/org/fuin/objects4j/jackson/Objects4JacksonUtils.java) utility class